:seedling: Refactor `ReviewFields`
Refactor the `ReviewFields` component:

  - The component is used on the Review tab of application and
    archetype details drawer

  - The input prop `reviews` was not used in any current containing
    component.  Push the `useFetchReviews()` hook down to the component
    itself.  This keeps the data fetch as close as possible to where
    it is used.

  - Move the component to `@app/components/detail-drawer` folder
    since it is used in more than one page grouping

Refactoring in advance of #1745
